requirements.txt
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
boto3==1.34.144
botocore==1.34.144
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorcet==3.1.0
colorlog==6.8.2
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dask==2024.7.0
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1719378673901/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
Deprecated==1.2.14
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1720869315914/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
figrid==0.1.6
flake8==7.1.0
fonttools==4.53.1
fsspec==2024.6.1
future==1.0.0
globus-sdk==3.42.0
graphtools==1.5.3
graphviz==0.20.3
ibl-neuropixel==1.1.3
iblatlas==0.5.3
ibllib==2.38.0
iblutil==1.11.0
idna==3.7
imagecodecs==2024.6.1
imageio==2.34.2
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1719361860083/work
importlib_resources==6.4.0
iniconfig==2.0.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1701831663892/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
jmespath==1.0.1
joblib==1.4.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1716472197302/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1710257447442/work
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.43.0
locket==1.0.0
lxml==5.2.2
matplotlib==3.9.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mccabe==0.7.0
mtscomp==1.0.2
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.2.1
npTDMS==1.9.0
nptyping==2.5.0
numba==0.60.0
numpy==1.26.4
ONE-api==2.8.1
opencv-python-headless==4.10.0.84
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==2.0.3
-e git+https://github.com/int-brain-lab/paper-reproducible-ephys.git@37fec5d07820a342790fd230a9f12c0ce85276c6#egg=paper_reproducible_ephys
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
partd==1.4.2
patsy==0.5.6
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
phate==1.0.11
phylib==2.6.0
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.4.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
pluggy==1.5.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1719274564771/work
Psychofit==1.0.0.post0
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyarrow==17.0.0
pycodestyle==2.12.0
pycparser==2.22
pyflakes==3.2.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
PyGSP==0.5.1
PyJWT==2.8.0
pynndescent==0.5.13
pynrrd==1.0.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
pytest==8.2.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytz==2024.1
PyYAML==6.0.1
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1715024408309/work
requests==2.32.3
s-gd2==1.8.1
s3transfer==0.10.2
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
scprep==1.2.3
seaborn==0.13.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
slidingRP==1.1.1
sobol-seq==0.2.0
sparse==0.15.4
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
statsmodels==0.14.2
svgutils==0.3.4
tasklogger==1.2.0
threadpoolctl==3.5.0
tifffile==2024.7.2
tomli==2.0.1
toolz==0.12.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1717722887379/work
tqdm==4.66.4
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.1
umap==0.1.1
umap-learn==0.5.6
urllib3==1.26.19
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
wrapt==1.16.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1718013267051/work
